Move slash command to dedicated package
