+++ backend/appBackend/src/jsMain/kotlin/com/gchristov/thecodinglove/JavascriptMain.kt
@@ -13,8 +13,8 @@ import com.gchristov.thecodinglove.search.SearchApiService
 import com.gchristov.thecodinglove.search.SearchModule
 import com.gchristov.thecodinglove.searchdata.SearchDataModule
 import com.gchristov.thecodinglove.slack.SlackModule
-import com.gchristov.thecodinglove.slack.SlackSlashCommandApiService
-import com.gchristov.thecodinglove.slack.SlackSlashCommandPubSubService
+import com.gchristov.thecodinglove.slack.slashcommand.SlackSlashCommandApiService
+import com.gchristov.thecodinglove.slack.slashcommand.SlackSlashCommandPubSubService
 import com.gchristov.thecodinglove.slackdata.SlackDataModule
 
 fun main() {
