+++ backend/slack/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/SlackModule.kt
@@ -5,10 +5,11 @@ import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSubSender
 import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSubServiceRegister
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
-import com.gchristov.thecodinglove.slack.usecase.RealVerifySlackRequestUseCase
+import com.gchristov.thecodinglove.slack.slashcommand.SlackSlashCommandApiService
+import com.gchristov.thecodinglove.slack.slashcommand.SlackSlashCommandPubSubService
 import com.gchristov.thecodinglove.slackdata.SlackRepository
 import com.gchristov.thecodinglove.slackdata.domain.SlackConfig
-import com.gchristov.thecodinglove.slackdata.usecase.VerifySlackRequestUseCase
+import com.gchristov.thecodinglove.slackdata.VerifySlackRequestUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.datetime.Clock
 import kotlinx.serialization.json.Json
