+++ backend/slack/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/slashcommand/SlackSlashCommandApiService.kt
@@ -1,4 +1,4 @@
-package com.gchristov.thecodinglove.slack
+package com.gchristov.thecodinglove.slack.slashcommand
 
 import arrow.core.Either
 import arrow.core.flatMap
@@ -8,10 +8,10 @@ import com.gchristov.thecodinglove.commonservicedata.api.*
 import com.gchristov.thecodinglove.commonservicedata.exports
 import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSubSender
 import com.gchristov.thecodinglove.commonservicedata.pubsub.sendMessage
+import com.gchristov.thecodinglove.slackdata.VerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slackdata.api.ApiSlackSlashCommand
 import com.gchristov.thecodinglove.slackdata.domain.SlackConfig
 import com.gchristov.thecodinglove.slackdata.domain.toPubSubMessage
-import com.gchristov.thecodinglove.slackdata.usecase.VerifySlackRequestUseCase
 import kotlinx.serialization.json.Json
 
 class SlackSlashCommandApiService(
