+++ backend/slack/src/jsTest/kotlin/com/gchristov/thecodinglove/slack/RealVerifySlackRequestUseCaseTest.kt
@@ -1,9 +1,9 @@
-package com.gchristov.thecodinglove.slack.usecase
+package com.gchristov.thecodinglove.slack
 
 import arrow.core.Either
 import com.gchristov.thecodinglove.kmpcommontest.FakeCoroutineDispatcher
+import com.gchristov.thecodinglove.slackdata.VerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slackdata.domain.SlackConfig
-import com.gchristov.thecodinglove.slackdata.usecase.VerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slacktestfixtures.FakeSlackApiRequest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestResult
